Bug #73759, fix heap dump corruption caused by duplicate message listeners which caused two heap dumps to be created and wrong heap dump length to be returned in the responses. (#2713)
Scheduler was registering message listeners even when embedded in server, causing duplicate responses to cluster messages. Only register listeners when running as standalone scheduler (ScheduleServer=true).
Also fix int overflow for heap dumps >2GB by using long for file sizes and decrease the buffer size to 100MB.

+++ core/src/main/java/inetsoft/web/admin/server/ServerMonitorMBean.java
@@ -96,7 +96,7 @@ public class ServerMonitorMBean {
     * @since 12.0
     */
    @ManagedOperation
-   public int getHeapDumpLength(String id) {
+   public long getHeapDumpLength(String id) {
       return serverService.getHeapDumpLength(id);
    }
 
@@ -113,7 +113,7 @@ public class ServerMonitorMBean {
     * @since 12.0
     */
    @ManagedOperation
-   public byte[] getHeapDumpContent(String id, int offset, int length) {
+   public byte[] getHeapDumpContent(String id, long offset, int length) {
       return serverService.getHeapDumpContent(id, offset, length);
    }
